backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[project]
name = "dagster-designer-backend"
version = "0.1.0"
description = "Backend API for Dagster visual pipeline designer"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.2",
    "pydantic-settings>=2.6.0",
    "dagster>=1.9.0",
    "dagster-dbt>=0.25.0",
    "dagster-fivetran>=0.25.0",
    "dagster-sling>=0.25.0",
    "dagster-dlt>=0.25.0",
    "gitpython>=3.1.43",
    "pyyaml>=6.0.2",
    "jinja2>=3.1.4",
    "python-multipart>=0.0.17",
    "aiofiles>=24.1.0",
    "toml>=0.10.2",
    "psutil>=5.9.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]
